backend/requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
pydantic[email]
pymongo

